Adds logic for console based tic tac toe game
In script.js, Cell(), GameBoard(), and GameController() are defined.

Cell() creates objects for each space on the tic tac toe board. Each
space can take on the values: 0, 1, or 2. 0 represents empty space,
1 or 2 represents player tokens.

GameBoard() sets up a 3x3 grid of Cell(). The GameBoard() keeps track
of all the moves that each player has performed.

GameController() controls the logic of the game. GameContoller() allows
the game to play rounds and eventually determine a winner (or tie).