Adds ability to update game board with ScreenController()
Each time the player clicks on a Cell on the GameBoard, the Cell will
update to reflect the token that is placed on that Cell.

Problems: currently, after the game has ended, PlayerDisplay will keep
changing, when the desired effect is that it only shows the relevant
game over message and does not change until the user clicks the Start
Game button to play again.