Fixes player-display update problem
Previously, the player-display would not display the end game message
correctly and would change to player's turn message when a player
clicks on the board even after the game has ended. This is now fixed
and board clicks after the game has ended does not make any changes
to the end. The user must click Start Game button to start a new game.