Adds ability to change player names
Index.html now shows each player's name, score, and button to change
the name. By default, the names are "Player One" and "Player Two". When
user clicks on one of the Change Name buttons, a prompt will ask
a string for name change. Once inputted, the name will change both
at the top and in Player-Display.